This is the moment.
This is the day,
When we send all our doubts and demons
On their way.
Every endeavor,
We have made - ever -
Is coming into play,
Is here and now - today.
This is the moment,
This is the time,
When the momentum and the moment
Are in rhyme!
Give us this moment -
This precious chance -
We'll gather up our past
And make some sense at last!
This is the moment,
When all we've done -
All the dreaming,
Scheming and screaming,
Become one!
This is the day -
See it sparkle and shine,
When all we've lived for
Becomes thine!
For all these years,
We've faced the world alone,
And now the time has come
To prove to them
We made it on our own!
This is the moment! -
Our final test! -
Destiny beckoned,
We never reckoned,
Second Best!
We won't look down,
We must not fall!
This is the moment,
The sweetest moment of them all!
This is the moment!
Damn all the odds!
This day, or never,
We'll sit forever
With the gods!
When we look back,
We will always recall,
Moment fo